requirements.txt
click==8.1.7            
nltk==3.8.1      
numpy==1.24.4         
pandas==2.0.3         
pytest==8.2.2      
python-dotenv==1.0.1      
requests==2.32.3   
tinydb==4.8.0      
tokenizers==0.19.1     
torch==2.3.1         
transformers==4.41.2     